Set default GOPATH to golang's default
The default GOPATH is '$HOME/go', but the bootstrap script was setting
it to '$HOME/gopath'. There shouldn't be a need to change this from the
language default location unless the user has already set their $GOPATH
environment variable to point to somewhere else, so this changes the
default to match.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>