Refactor API code to make more modular
This refactors the API code to follow the beego recommendations and to
make it more modular. This restruture should make things a little more
manageable as we add file and object support, and it is needed to be
able to support alternative northbound APIs (Swordfish) for accessing
the some functionality but routed through a different endpoint.

Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>